# Speed up make_holiday_features by up to 35%